dataflow_example/requirements.txt
apache-beam[gcp]==2.46.0
google-cloud-pubsub==2.18.0
google-cloud-bigquery==3.11.0
google-cloud-storage==2.10.0